+++ config/config_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func TestReadConfig(t *testing.T) {
-	path := "../testData/config/bitxhub.toml"
+	path := "../testdata/config/bitxhub.toml"
 	config, err := UnmarshalConfig("", path)
 	assert.Nil(t, err)
 	assert.Equal(t, 4, len(config.Addrs))
