+++ rpc_test.go
@@ -75,7 +75,7 @@ func TestEthRPC_EthCall(t *testing.T) {
 	time.Sleep(time.Second)
 	contractAbi, err := utils.LoadAbi("./testdata/storage.abi")
 	require.Nil(t, err)
-	args, err := Encode(&contractAbi, "store", "5")
+	args, err := utils.Decode(&contractAbi, "store", "5")
 	require.Nil(t, err)
 	_, err = client.Invoke(account.PrivateKey, &contractAbi, addresses[0], "store", args)
 	require.Nil(t, err)
@@ -102,7 +102,7 @@ func TestInvokeEthContract(t *testing.T) {
 	time.Sleep(time.Second)
 	contractAbi, err := utils.LoadAbi("./testdata/storage.abi")
 	require.Nil(t, err)
-	args, err := Encode(&contractAbi, "store", "2")
+	args, err := utils.Decode(&contractAbi, "store", "2")
 	require.Nil(t, err)
 	_, err = client.Invoke(account.PrivateKey, &contractAbi, addresses[0], "store", args)
 	require.Nil(t, err)
@@ -381,7 +381,7 @@ func prepareContract(t *testing.T, cli Client, privateKey *ecdsa.PrivateKey, con
 	}
 	bodyBytes1, err := json.Marshal(orgInput)
 	assert.Nil(t, err)
-	input1, err := utils.Decode(abiEvent, "registerOrg", bodyBytes1)
+	input1, err := utils.DecodeBytes(abiEvent, "registerOrg", bodyBytes1)
 	require.Nil(t, err)
 	_, err = cli.Invoke(privateKey, &contractAbi, contractAddr, "registerOrg", input1)
 	assert.Nil(t, err)
@@ -397,7 +397,7 @@ func prepareContract(t *testing.T, cli Client, privateKey *ecdsa.PrivateKey, con
 	}
 	bodyBytes2, err := json.Marshal(userInstance)
 	assert.Nil(t, err)
-	input2, err := utils.Decode(abiEvent, "registerUser", bodyBytes2)
+	input2, err := utils.DecodeBytes(abiEvent, "registerUser", bodyBytes2)
 	require.Nil(t, err)
 	_, err = cli.Invoke(privateKey, &contractAbi, contractAddr, "registerUser", input2)
 	assert.Nil(t, err)
@@ -455,7 +455,7 @@ func TestInvokeTupleContract(t *testing.T) {
 
 	bodyBytes, err := json.Marshal(input)
 	assert.Nil(t, err)
-	args, err := utils.Decode(abiEvent, "publish", bodyBytes)
+	args, err := utils.DecodeBytes(abiEvent, "publish", bodyBytes)
 	assert.Nil(t, err)
 	_, err = cli.Invoke(account.PrivateKey, &contractAbi, string(code), "publish", args)
 	assert.Nil(t, err)
